tools/api-key-manager/requirements.txt
textual>=0.47.0
httpx>=0.27.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
